+++ src/ext_widgets/scroll_view/scroll_view.c
@@ -368,6 +368,7 @@ static ret_t scroll_view_on_event(widget_t* widget, event_t* e) {
 
 static ret_t scroll_view_on_paint_children(widget_t* widget, canvas_t* c) {
   rect_t r_save;
+  vgcanvas_t* vg = canvas_get_vgcanvas(c);
   rect_t r = rect_init(c->ox, c->oy, widget->w, widget->h);
 
   scroll_view_t* scroll_view = SCROLL_VIEW(widget);
@@ -378,6 +379,12 @@ static ret_t scroll_view_on_paint_children(widget_t* widget, canvas_t* c) {
   canvas_get_clip_rect(c, &r_save);
 
   r = rect_intersect(&r, &r_save);
+
+  if (vg != NULL) {
+    vgcanvas_save(vg);
+    vgcanvas_clip_rect(vg, (float_t)r.x, (float_t)r.y, (float_t)r.w, (float_t)r.h);
+  }
+
   canvas_set_clip_rect(c, &r);
   if (scroll_view->on_paint_children) {
     scroll_view->on_paint_children(widget, c);
@@ -387,6 +394,11 @@ static ret_t scroll_view_on_paint_children(widget_t* widget, canvas_t* c) {
   canvas_set_clip_rect(c, &r_save);
   canvas_untranslate(c, xoffset, yoffset);
 
+  if (vg != NULL) {
+    vgcanvas_clip_rect(vg, (float_t)r_save.x, (float_t)r_save.y, (float_t)r_save.w,
+                       (float_t)r_save.h);
+    vgcanvas_restore(vg);
+  }
   return RET_OK;
 }